Check dates instead of years
The municipality filtering function now implicitly assumes that the year supplied covers beginning and end of year. This is to match the logic behind the get_klass function, and the fact that all municipality changes happen at the beginning of a year.

The get_municipality structure function now queries the last day of the year when getting the muni structure from get_klass.

Note that the functions now heavily assume that municipality IDs are valid for entire years.